feat: 3rd-party-friendly test-mediator + self-loopback routing + types relocation
Three loosely-related improvements to make running mediator/test-mediator
from outside this workspace ergonomic and to fix the underlying routing-2.0
self-forward bug. Each change is independently reviewable but they ship
together so the version bumps line up.

## test-mediator (0.1.0 → 0.2.0)

- `TestMediatorBuilder::enable_external_forwarding(bool)` — surface the
  existing `ForwardingConfig::external_forwarding` knob so a test can
  collapse all forwards to local delivery without overriding the whole
  config struct.
- `TestMediatorHandle::register_local_did(&self, did)` and
  `TestMediatorHandle::add_user(alias)` — runtime counterparts to the
  existing builder-time `local_did` setter, removing the
  chicken-and-egg between "user DIDs need the mediator's DID" and
  "mediator DID isn't generated until spawn".
- `TestMediator::with_users(["alice", "bob"])` — convenience that spawns
  the mediator and returns pre-registered users whose service URI is
  the mediator's DID (the routing-2.0 shape, not the HTTP URL).
- `TestEnvironment::add_user`: switch user DID's service URI from the
  mediator's HTTP URL to its DID. The HTTP-URL shape silently broke
  forwarding tests (every `routing/2.0/forward` looked remote and
  pushed to FORWARD_Q).
- README: new "Local vs. remote routing" section documenting the
  service-endpoint-shape rule + the new helpers.

## mediator (0.14.1 → 0.15.0)

- `service_endpoint_for_remote` (routing.rs) now treats a service URI
  as local when its `(host, port)` matches the mediator's bound address
  or any operator-declared alias. URLs that round-trip back to this
  instance via a different hostname or behind a load balancer no longer
  land in FORWARD_Q.
- New `[server.local_endpoints]` config + `MediatorBuilder::local_endpoints`
  for declaring those aliases. `listen_address` is always treated as
  local, so default deployments (and the test-mediator) need nothing.
- `SharedData::self_authorities` caches the parsed `(host, port)` set
  at startup.
- 8 unit tests covering authority matching: case-insensitive host,
  default ports for http/https/ws/wss, host/port mismatches.

## sdk (0.17.0 → 0.18.0) + mediator-common (0.13.0 → 0.14.0)

The mediator's storage trait used to import `MediatorACLSet`, `Account`,
`Folder`, `ProblemReport`, etc. from `affinidi-messaging-sdk` — wrong
direction; storage backends shouldn't depend on the client SDK. Moved
the protocol-vocabulary types (ACLs, accounts, ACL-handler responses,
admin types, message-pickup types, problem reports) into
`affinidi-messaging-mediator-common::types::*` and re-exported from the
SDK at their original paths. `affinidi-messaging-sdk` now depends on
`affinidi-messaging-mediator-common`, not vice versa.

Breaking change: `MediatorACLSet::*` methods that previously returned
`Result<_, ATMError>` now return `Result<_, ACLError>` — a lightweight
error type local to common. `From<ACLError> for ATMError` is provided,
so callers using `?` against `ATMError` are unaffected. Direct match-
on-variant callers were updated (one site in `affinidi-messaging-didcomm-service`).

## Version bumps

Workspace caret-pin policy: only consumers using `version = "0.13"` (now
"0.14") or `version = "0.17"` (now "0.18") were updated.

- affinidi-messaging-mediator-common: 0.13.0 → 0.14.0
- affinidi-messaging-sdk: 0.17.0 → 0.18.0
- affinidi-messaging-mediator: 0.14.1 → 0.15.0
- affinidi-messaging-test-mediator: 0.1.0 → 0.2.0
