+++ crates/messaging/affinidi-messaging-helpers/Cargo.toml
@@ -21,7 +21,7 @@ path = "src/mediator_administration/main.rs"
 [dependencies]
 # ── Affinidi Crates ──────────────────────────────────────────────────────
 ## SDK for connecting to and communicating with a running mediator
-affinidi-messaging-sdk = { path = "../affinidi-messaging-sdk", version = "0.17" }
+affinidi-messaging-sdk = { path = "../affinidi-messaging-sdk", version = "0.18" }
 ## TDK for DID resolution and crypto utilities
 affinidi-tdk = { path = "../../tdk/affinidi-tdk", version = "0.7" }
 ## DIDComm message types — used by example binaries
