+++ crates/messaging/affinidi-messaging-mediator/affinidi-messaging-mediator-common/src/database/mod.rs
@@ -1,5 +1,5 @@
 use crate::errors::MediatorError;
-use affinidi_messaging_sdk::messages::problem_report::{ProblemReportScope, ProblemReportSorter};
+use crate::types::problem_report::{ProblemReportScope, ProblemReportSorter};
 use axum::http::StatusCode;
 use config::DatabaseConfig;
 use redis::AsyncConnectionConfig;
