+++ crates/messaging/affinidi-messaging-mediator/affinidi-messaging-mediator-common/src/store/redis/database/get.rs
@@ -1,6 +1,6 @@
 use super::Database;
 use crate::errors::MediatorError;
-use affinidi_messaging_sdk::messages::MessageListElement;
+use crate::types::messages::MessageListElement;
 use itertools::Itertools;
 use redis::{Value, from_redis_value};
 use subtle::ConstantTimeEq;
